+++ src/echo/__init__.py
@@ -1,3 +1,3 @@
 """Echo programming language."""
 
-__version__ = "0.6.7"
+__version__ = "0.6.8"